project root/requirements.txt
flask>=2.0.1
mysql-connector-python>=8.0.28
pytesseract>=0.3.8
Pillow>=9.0.0
PyPDF2>=3.0.1
PyMuPDF>=1.19.6
werkzeug>=2.0.1 